chore: update useAppendDomNode.ts
